Add files via upload
In this portfolio, I've included a range of web development projects I've worked on, such as a React-based weather app and a simple e-commerce site made with Ruby on Rails and HTML, CSS, and JavaScript. Code examples and thorough documentation are provided for every project. We always value your opinions!